Issue-159 - Extract connect_to_app
